Move greetings.py commands into handmade_bonus.py
